新增  CUDA 檢查、試聽按鈕 修復 聲音播放問題
三項新功能：
  - CUDA 檢查：「進階」頁運算裝置旁的「檢查」按鈕 → 背景執行緒檢查（不凍結 UI、防重入）→ 彈出診斷對話框（torch/CUDA/cuDNN 版本、GPU 型號與 VRAM、四種狀態說明、建議裝置值）。另含
  fallback：選 cuda 但初始化失敗時自動退回 CPU 繼續偵測，托盤通知提示。
  - 聲音修復：底層全面改用 QMediaPlayer＋QAudioOutput，mp3/wav 都能播；popup 關窗一併停音；解碼失敗記 log 不崩潰。
  - 試聽按鈕：兩個音檔欄位「輸入框＋瀏覽＋🔊試聽」同列，可播可停、未儲存即可試聽、互斥單播、關窗自停、無效檔顯示紅字錯誤。

  17 項錯誤全數修正（暫停狀態機死鎖、斷線偵測、ROI 錯位、設定互蓋、無效的紀錄開關……）。

  對抗審查再抓出 11 個新問題並已修正，包括一個 high：ROI 框選原本可寫入越界值，配上新加的 config 驗證會讓程式下次啟動把自己鎖死——已加 clamp＋舊值載入遷移。審查也補上了：YAML
  語法錯誤的友善訊息、CUDA 遲到結果的請求綁定、低 FPS 來源的健康判定寬限期等。

+++ src/types.py
@@ -79,7 +79,8 @@ class TimerSnapshot:
 
 @dataclass(frozen=True)
 class ReminderContext:
-    work_minutes: int
+    # §4.14: float — keep fractional minutes (e.g. 0.1 or 25.5) intact for countdowns.
+    work_minutes: float
     media_path: str
     media_type: str
     sound_path: str
